Use command line interface as base for simple console application template
