+++ src/consolebox/src/CommandBox/CommandBox.csproj
@@ -18,31 +18,26 @@
 		<!--#endif-->
 		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.1" />
 		<PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.1" />
+		<PackageReference Include="Microsoft.Extensions.FileProviders.Embedded" Version="8.0.10" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.1" />
 		<!--#if (serilog)-->
 		<PackageReference Include="Serilog.Extensions.Logging" Version="8.0.0" />
 		<PackageReference Include="Serilog.Sinks.Console" Version="6.0.0" />
 		<PackageReference Include="Serilog.Enrichers.Environment" Version="3.0.1" />
-		<!--#if (!basic)-->
 		<PackageReference Include="Serilog.Sinks.File" Version="6.0.0" />
 		<PackageReference Include="Serilog.Settings.Configuration" Version="8.0.4" />
-		<!--#endif-->
 		<!--#else-->
 		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.0.0" />
 		<!--#endif-->
-		<!--#if (!basic)-->
 		<PackageReference Include="CommandLineParser" Version="2.9.1" />
-		<!--#if (settings)-->
-		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="8.0.2" />
-		<!--#endif-->
-		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.1" />
-		<PackageReference Include="Microsoft.Extensions.FileProviders.Embedded" Version="8.0.10" />
+		<!--#if (options)-->
+		<PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0" />
 		<!--#endif-->
 		<!--#if (nswag)-->
 		<PackageReference Include="Microsoft.Extensions.Http" Version="8.0.1" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
 		<!--#endif-->
 	</ItemGroup>
-	<!--#if (!basic)-->
 	<ItemGroup>
 		<EmbeddedResource Include="appsettings.json">
 			<CopyToOutputDirectory>Never</CopyToOutputDirectory>
@@ -51,7 +46,6 @@
 	<ItemGroup>
 		<None Remove="appsettings.json" />
 	</ItemGroup>
-	<!--#endif-->
 	<!--#if (nswag)-->	
 	<Target Name="NSwag" BeforeTargets="BeforeBuild" Condition="!Exists('API_CLIENT.cs')">
 		<PropertyGroup>
