+++ src/consolebox/src/CommandBox/Options.cs
@@ -5,7 +5,7 @@ internal sealed class Options
     [Verb("run", isDefault: true, HelpText = "Run program.")]
     internal class Run
     {
-#if (settings)
+#if (options)
         [Option('p', "path", HelpText = "Directory path.", Required = true)]
         public string Path { get; set; }
 #endif
