+++ src/consolebox/src/CommandBox/Options/SettingsOptions.cs
@@ -1,0 +1,4 @@
+﻿internal class SettingsOptions
+{
+    public string MessageFormat { get; set; }
+}
