+++ src/consolebox/src/CommandBox/appsettings.json
@@ -21,11 +21,13 @@
         }
       }
     ]
+  //#endif  
+  //#if (options)
   },
-  //#endif
-  //#if (settings)
-  "ConsoleBox": {
-    "Message": "Hello, World!\nDirectory {path:l} contains {file} files"
+  "Settings": {
+    "MessageFormat": "Hello, World!\nDirectory {path:l} contains {file} files"
+  }
+  //#else
   }
   //#endif
 }
