--- src/consolebox/src/ConsoleBox/Settings.cs
@@ -1,4 +1,0 @@
-﻿internal class Settings
-{
-    public string Message { get; set; }
-}
