+++ src/nugetbox/.template.config/template.json
@@ -2,7 +2,7 @@
   "$schema": "http://json.schemastore.org/template",
   "author": "Matt Janda",
   "classifications": ["JandaBox/NuGet"],
-  "name": "JandaBox NuGet Class Library",
+  "name": "JandaBox NuGet Class Library Solution",
   "generatorVersions": "[1.0.0.0-*)",
   "description": "JandaBox nugetbox provides NuGet ready class library template with github actions and semantic versioning.",
   "groupIdentity": "nugetbox",
